requirements.txt
# Libraries
absl-py # FLAGS etc; like internal
watchdog # local file system events; for simulator only
decorator # TODO: why is this a library?
lucid==0.0.8 # for io - should be spun off or heavily redesigned?
toposort # because we're writing a real language with dependency resolution \o/
utilspie # just for frozendict... O_o

# Google API
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-datastore==1.6.0
google-cloud-storage
google-cloud-pubsub==0.32.1

# Testing
pytest
pytest-watch
pytest-testmon
pytest-mock
mypy